Resolved the issue with the ball rolling on the roof/floor every collision
When a collision occurs the ball's velocity automatically gets overwritten by the physics engine in such a way that makes it slide along the roof/wall.

Because the "new" velocity is sent with a y direction of 0 it would assume the angle of impact was incredibly shallow, causing the rolling behaviour.
Now we store the velocity before impact and use that when calculating the new impact velocity.