Added hacks to fix some issues.
The paddle x location is now constantly fixed. This should prevent the issue where the paddles would move on collision with the ball.
This is likely due to the collision boxes overlapping by a bit, forcing the physics engine to separate the two objects by moving them in the directions opposite the collision vector.

The ball now correctly re-spawns in the middle after scoring a point. This was done with a hack that causes the set_global_position to be called on a 0 second timer, this hacks around the issue of the ball disappearing.
The underlying cause is likely a race condition of some kind. Perhaps the physics engine sees the change in global position from the previous frame and infers that a massive acceleration is required for that movement, and thus assumes the velocity and
launches it out of the park? Don't have complete answers at the moment.

These are temporary fixes and the underlying issue should be looked at instead of relying on these hacks.